requirements.txt
amqp==5.2.0
asgiref==3.8.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssutils==2.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
django-inlinecss==0.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
future==1.0.0
gunicorn==21.2.0
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pynliner==0.8.0
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
